requirements.txt
pdfplumber>=0.10.0
pymupdf>=1.23.0
Pillow>=10.0.0
imagehash>=4.3.0
numpy>=1.24.0
